Use int32 instead of int type

+++ noarch/string.go
@@ -98,15 +98,15 @@ func Strchr(str []byte, ch int32) []byte {
 	return nil
 }
 
-func Memset(dst interface{}, val int, size int) interface{} {
+func Memset(dst interface{}, val int32, size int32) interface{} {
 	vDst := reflect.ValueOf(dst).Type()
 	switch vDst.Kind() {
 	case reflect.Slice, reflect.Array:
 		vDst = vDst.Elem()
 	}
-	baseSizeDst := int(vDst.Size())
-	data := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(dst, baseSizeDst, 1)))
-	var i int
+	baseSizeDst := int32(vDst.Size())
+	data := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(dst, baseSizeDst, int32(1))))
+	var i int32
 	var vb = byte(val)
 	for i = 0; i < size; i++ {
 		data[i] = vb
@@ -114,22 +114,22 @@ func Memset(dst interface{}, val int, size int) interface{} {
 	return dst
 }
 
-func Memcpy(dst interface{}, src interface{}, size int) interface{} {
+func Memcpy(dst interface{}, src interface{}, size int32) interface{} {
 	vDst := reflect.ValueOf(dst).Type()
 	switch vDst.Kind() {
 	case reflect.Slice, reflect.Array:
 		vDst = vDst.Elem()
 	}
-	baseSizeDst := int(vDst.Size())
+	baseSizeDst := int32(vDst.Size())
 	vSrc := reflect.ValueOf(src).Type()
 	switch vSrc.Kind() {
 	case reflect.Slice, reflect.Array:
 		vSrc = vSrc.Elem()
 	}
-	baseSizeSrc := int(vSrc.Size())
-	bDst := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(dst, baseSizeDst, 1)))
-	bSrc := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(src, baseSizeSrc, 1)))
-	var i int
+	baseSizeSrc := int32(vSrc.Size())
+	bDst := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(dst, baseSizeDst, int32(1))))
+	bSrc := *(*[]byte)(unsafe.Pointer(UnsafeSliceToSlice(src, baseSizeSrc, int32(1))))
+	var i int32
 	for i = 0; i < size; i++ {
 		bDst[i] = bSrc[i]
 	}
